since recent changes were apparently untested and resulted in syntax and nil errors, reverting the deploy script

+++ bin/deploy.rb
@@ -1,6 +1,5 @@
 def rmdir_r(dir)
   Dir.chdir(dir)
-  
   Dir['**/{*,.*}'].select do |f|
     n = f.split('/').last
     n != '..' and n != '.'
@@ -11,45 +10,42 @@ def rmdir_r(dir)
       File.delete(f)
     end
   end
-  
   Dir.delete(dir)
 end
 
 def main
   dotfiles = File.join(Dir.home, '.dotfiles')
   entries = Dir.entries(dotfiles).select { |f| f[0] != '.' }
-  
   existing = entries.select do |f|
     p = File.join(Dir.home, '.' + f)
     File.exists?(p) and !File.symlink?(p)
   end
-  
+  response = 'y'
   if !existing.empty?
     puts 'WARNING: This will overwrite the following dotfiles:'
-    
     existing.each do |f|
       puts (File.directory?(f) ? 'd ' : 'f ') + f
     end
-    
     print 'Are you sure you want to overwrite these files? (y/n) '
-    response ||= gets
+    response = gets
   end
-  
   if !entries.empty? and response[0].downcase == 'y'
-    entries.reject!{ |f| ['bin', 'README.md'].include?(f) || f[0] == '.' }.each do |f|
-      p = File.join(Dir.home, '.' + f)
-      if !File.symlink?(p)
-        if File.exists?(p)
-          if File.directory?(p)
-            rmdir_r(p)
+    entries.each do |f|
+      if f != 'bin' and f[0] != '.' and f != 'README.md'
+        p = File.join(Dir.home, '.' + f)
+        if !File.symlink?(p)
+          if File.exists?(p)
+            if File.directory?(p)
+              rmdir_r(p)
+            else
+              File.delete(p)
+            end
+            puts (File.directory?(f) ? 'd ' : 'f ') + f + ' (overwrote)'
           else
-            File.delete(p)
+            puts (File.directory?(f) ? 'd ' : 'f ') + f + ' (wrote)'
           end
-          puts (File.directory?(f) ? 'd ' : 'f ') + f + ' (overwrote)'
-        else
-          puts (File.directory?(f) ? 'd ' : 'f ') + f + ' (wrote)'
+          File.symlink(File.join(dotfiles, f), p)
         end
-        File.symlink(File.join(dotfiles, f), p)
       end
     end
   else
@@ -61,3 +57,4 @@ end
 if __FILE__ == $0
   main
 end
+
